+++ src/Validators/Card/DuplicateCardValidator.php
@@ -78,8 +78,8 @@ class DuplicateCardValidator
     private function validateDocumentNumber(): void
     {
         $documentNumber = $this->duplicateCard->documentNumber;
-        if (empty($documentNumber) || ! is_string($documentNumber) || ! is_numeric($documentNumber)) {
-            throw new InvalidArgumentException('document number should be a numeric string');
+        if (empty($documentNumber) || ! is_string($documentNumber) || ! preg_match('/^[0-9A-Za-z]+$/', $documentNumber)) {
+            throw new InvalidArgumentException('document number should be an alphanumeric string');
         }
 
         $documentValidator = new CpfCnpjValidator($documentNumber);