+++ src/Validators/Pix/BankAccountValidator.php
@@ -64,8 +64,8 @@ class BankAccountValidator
     private function validateDocumentNumber(): void
     {
         $documentNumber = $this->bankAccount->documentNumber;
-        if (empty($documentNumber) || ! is_string($documentNumber) || ! is_numeric($documentNumber)) {
-            throw new InvalidArgumentException('document number should be a numeric string');
+        if (empty($documentNumber) || ! is_string($documentNumber) || ! preg_match('/^[0-9A-Za-z]+$/', $documentNumber)) {
+            throw new InvalidArgumentException('document number should be an alphanumeric string');
         }
     }